% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_col_names.R
\name{mg_col_names}
\alias{mg_col_names}
\title{Return column names of a mongo DB table (collection).}
\usage{
mg_col_names(table)
}
\arguments{
\item{table}{name (as character).}
}
\value{
character
}
\description{
Return column names of a mongo DB table (collection).
}
