% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_column_exists.R
\name{mg_column_exists}
\alias{mg_column_exists}
\title{Check if a table (collection) exists in a mongo DB.}
\usage{
mg_column_exists(table, name)
}
\arguments{
\item{table}{name as a character}

\item{name}{of the column as a character}
}
\value{
logical
}
\description{
Check if a table (collection) exists in a mongo DB.
}
