% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_create_new_column.R
\name{mg_create_new_column}
\alias{mg_create_new_column}
\title{Creates a new column in a mongo DB table.}
\usage{
mg_create_new_column(table, name, value)
}
\arguments{
\item{table}{name of the table.}

\item{name}{name of the new colum.}

\item{value}{default value for the new column.}
}
\value{
TRUE (if it doesn't error)
}
\description{
Creates a new column in a mongo DB table.
}
