% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_list_tables.R
\name{mg_list_tables}
\alias{mg_list_tables}
\title{List all tables in a mongo DB.}
\usage{
mg_list_tables()
}
\value{
character containing all table names in the DB.
}
\description{
List all tables in a mongo DB.
}
