% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_rm_column.R
\name{mg_rm_column}
\alias{mg_rm_column}
\title{Remove a column from a mongo DB table (collection).}
\usage{
mg_rm_column(table, name)
}
\arguments{
\item{table}{name as a character}

\item{name}{of the column as a character}
}
\value{
logical
}
\description{
Remove a column from a mongo DB table (collection).
}
