% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_select2.R
\name{mg_select2}
\alias{mg_select2}
\title{Select columns from a mongo DB table}
\usage{
mg_select2(
  table,
  ...,
  .query = "{}",
  .sort = "{}",
  .skip = 0,
  .limit = 0,
  .handler = NULL,
  .pagesize = 1000,
  ._id = FALSE
)
}
\arguments{
\item{table}{name of the table/collection (as character).}

\item{...}{column names (as symbols)}

\item{.query}{todo}

\item{.sort}{todo}

\item{.skip}{todo}

\item{.limit}{todo}

\item{.handler}{todo}

\item{.pagesize}{todo}

\item{._id}{Should the _id column be returned?}
}
\value{
A data frame with the selected columns
}
\description{
Select columns from a mongo DB table
}
