% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_table_exists.R
\name{mg_table_exists}
\alias{mg_table_exists}
\title{Check if a table (collection) exists in a mongo DB.}
\usage{
mg_table_exists(table)
}
\arguments{
\item{table}{name as a character}
}
\value{
logical
}
\description{
Check if a table (collection) exists in a mongo DB.
}
